requirements.txt
python-dotenv==1.0.1
openai==1.59.7
requests==2.31.0
streamlit==1.33.0